+++ Ascension/src/blockfighter/server/entities/buff/BuffXSpeedIncrease.java
@@ -1,0 +1,6 @@
+package blockfighter.server.entities.buff;
+
+public interface BuffXSpeedIncrease {
+
+    public double getXSpeedIncrease();
+}