+++ Ascension/src/blockfighter/shared/Globals.java
@@ -1,6 +1,5 @@
 package blockfighter.shared;
 
-import blockfighter.client.entities.particles.skills.other.ParticleHit;
 import blockfighter.client.entities.emotes.*;
 import blockfighter.client.entities.particles.*;
 import blockfighter.client.entities.particles.skills.bow.*;
@@ -247,12 +246,10 @@ public class Globals {
         UTILITY_DASH_BUFF_EMITTER((byte) 0x32, null, 0, ParticleUtilityDashBuffEmitter.class, PARTICLE_PARAM_PLAYER),
         UTILITY_DASH_BUFF_PARTICLE((byte) 0x33, "dashbuff", 1, ParticleUtilityDashBuffParticle.class, PARTICLE_PARAM_POS_AND_FACING),
         UTILITY_DASH_EMITTER((byte) 0x34, null, 0, ParticleUtilityDashEmitter.class, PARTICLE_PARAM_FACING_AND_PLAYER),
-        UTILITY_FORTIFY((byte) 0x35, "fortify", 20, ParticleUtilityFortify.class, PARTICLE_PARAM_PLAYER),
-        UTILITY_FORTIFY_BUFF((byte) 0x36, "fortifybuff", 8, ParticleUtilityFortifyBuff.class, PARTICLE_PARAM_POS_AND_FACING),
-        UTILITY_FORTIFY_BUFF_EMITTER((byte) 0x37, null, 0, ParticleUtilityFortifyEmitter.class, PARTICLE_PARAM_PLAYER),
-        UTILITY_CLONE_EMITTER((byte) 0x38, null, 0, ParticleUtilityCloneEmitter.class, PARTICLE_PARAM_PLAYER),
-        UTILITY_CLONE_PARTICLE((byte) 0x39, null, 0, ParticleUtilityCloneParticle.class, null),
-        HIT((byte) 0x3A, "hit", 6, ParticleHit.class, PARTICLE_PARAM_POS_AND_FACING);
+        UTILITY_ADRENALINE((byte) 0x35, null, 0, ParticleUtilityAdrenaline.class, PARTICLE_PARAM_PLAYER),
+        UTILITY_ADRENALINE_CLONE_EMITTER((byte) 0x36, null, 0, ParticleUtilityAdrenalineCloneEmitter.class, PARTICLE_PARAM_PLAYER),
+        UTILITY_ADRENALINE_CLONE_PARTICLE((byte) 0x37, null, 0, ParticleUtilityAdrenalineCloneParticle.class, null),
+        HIT((byte) 0x38, "hit", 6, ParticleHit.class, PARTICLE_PARAM_POS_AND_FACING);
 
         private final byte particleCode;
 
@@ -551,7 +548,7 @@ public class Globals {
             BOW_FROST = 0x09,
             BOW_STORM = 0x0A,
             BOW_VOLLEY = 0x0B,
-            UTILITY_FORTIFY = 0x0C,
+            UTILITY_ADRENALINE = 0x0C,
             SHIELD_ROAR = 0x0D,
             SHIELD_CHARGE = 0x0E,
             SHIELD_REFLECT = 0x0F,